# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TextMesh API
  description: |
    TextMesh is a text-only social media platform designed for scale.

    ## Authentication
    Most endpoints require authentication via Bearer token. Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - Standard endpoints: 100 requests/minute
    - Auth endpoints: 10 requests/minute

    ## Pagination
    All list endpoints support cursor-based pagination with `cursor` and `limit` parameters.
  version: 1.0.0
  contact:
    name: TextMesh API Support
    email: api@textmesh.io
    url: https://developers.textmesh.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.textmesh.io/v1
    description: Production
  - url: https://api.staging.textmesh.io/v1
    description: Staging
  - url: http://localhost:3000
    description: Development
...
security:
  - BearerAuth: []
paths:
...
  /posts:
    post:
      tags: [Posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /posts/{postId}:
    get:
      tags: [Posts]
      summary: Get a post by ID
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Posts]
      summary: Delete a post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Post deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /posts/{postId}/like:
    post:
      tags: [Posts]
      summary: Like a post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Post liked

    delete:
      tags: [Posts]
      summary: Unlike a post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Post unliked
...
  /posts/{postId}/bookmark:
    post:
      tags: [Posts]
      summary: Bookmark a post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Post bookmarked

    delete:
      tags: [Posts]
      summary: Remove bookmark
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Bookmark removed
  /posts/{postId}/replies:
    get:
      tags: [Posts]
      summary: Get replies to a post
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of replies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'

    post:
      tags: [Posts]
      summary: Reply to a post
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Reply created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
  /users/{userId}/posts:
    get:
      tags: [Posts]
      summary: Get user's posts
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: includeReplies
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: User's posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
        bio:
          type: string
        isVerified:
          type: boolean
        followerCount:
          type: integer
        followingCount:
          type: integer
        postCount:
          type: integer
        isFollowing:
          type: boolean
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        visibility:
          type: string
          enum: [PUBLIC, FOLLOWERS, PRIVATE]
        user:
          $ref: '#/components/schemas/User'
        parentId:
          type: string
          format: uuid
          nullable: true
        quotedPostId:
          type: string
          format: uuid
          nullable: true
        quotedPost:
          $ref: '#/components/schemas/Post'
        groupId:
          type: string
          format: uuid
          nullable: true
        likeCount:
          type: integer
        replyCount:
          type: integer
        repostCount:
          type: integer
        isLiked:
          type: boolean
        isReposted:
          type: boolean
        isBookmarked:
          type: boolean
        hashtags:
          type: array
          items:
            type: string
        mentions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    CreatePostRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 500
        visibility:
          type: string
          enum: [PUBLIC, FOLLOWERS, PRIVATE]
          default: PUBLIC
        groupId:
          type: string
          format: uuid
    PostResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Post'
    PostListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Post'
            nextCursor:
              type: string
              nullable: true
            hasMore:
              type: boolean

    # Groups
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object